requirements.txt
feedparser
anthropic
sqlalchemy
fastapi
uvicorn[standard]
apscheduler
python-dotenv
httpx
pytest
